CMakeLists.txt: Do not unconditionally use ccache
This causes build failures in Gentoo because we don't allow access to
ccache files if ccache is not enabled for build.
Fix this by adding a WITH_CCACHE cmake option and change behavior so that
cmake fails if WITH_CCACHE is enabled but ccache program cannot be found.

Gentoo-bug: https://bugs.gentoo.org/704560
Signed-off-by: Lars Wendler <polynomial-c@gentoo.org>